Added Tracking
I added some Google Analytics Tracking, with events. It will capture
each time the addon is activated and when it loads a new list of songs.
It will also track what songs are correct. I plan to make a setting that
will disable tracking, for people that does not like this.